move keymap to a separate lib
