+++ apps/frontend/src/app/Components/StatDisplay.tsx
@@ -1,6 +1,6 @@
 import { iconInlineProps } from '@genshin-optimizer/common/svgicons'
+import type { StatKey } from '@genshin-optimizer/gi/keymap'
 import { Box } from '@mui/material'
-import type { StatKey } from '../KeyMap'
 import KeyMap from '../KeyMap'
 import StatIcon from '../KeyMap/StatIcon'
 import ColorText from './ColoredText'
