+++ apps/frontend/src/app/PageCharacter/CharacterDisplay/Tabs/TabOptimize/Components/ChartCard/CustomTooltip.tsx
@@ -2,6 +2,7 @@ import { valueString } from '@genshin-optimizer/common/util'
 import { allArtifactSlotKeys } from '@genshin-optimizer/gi/consts'
 import type { ICachedArtifact } from '@genshin-optimizer/gi/db'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
+import type { Unit } from '@genshin-optimizer/gi/keymap'
 import {
   Button,
   CardContent,
@@ -21,7 +22,6 @@ import CloseButton from '../../../../../../Components/CloseButton'
 import SqBadge from '../../../../../../Components/SqBadge'
 import { DataContext } from '../../../../../../Context/DataContext'
 import { input } from '../../../../../../Formula'
-import type { Unit } from '../../../../../../KeyMap'
 import { ArtifactSetBadges } from '../ArtifactSetBadges'
 import type EnhancedPoint from './EnhancedPoint'
