+++ libs/gi/keymap/README.md
@@ -1,0 +1,6 @@
+# gi-keymap
+
+This library was generated with [Nx](https://nx.dev).
+
+This hosts some key-> english for displaying formula related content for the Waverider engine.
+Also has some numbers and multipliers as well.
