+++ libs/gi/keymap/project.json
@@ -1,0 +1,22 @@
+{
+  "name": "gi-keymap",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/gi/keymap/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/gi/keymap/**/*.ts"]
+      }
+    },
+    "test": {
+      "executor": "@nx/vite:test",
+      "options": {
+        "config": "libs/gi/keymap/vitest.config.ts"
+      }
+    }
+  },
+  "tags": []
+}
